backend/requirements.txt
annotated-types==0.7.0
anyio==4.10.0
asgiref==3.8.1
certifi==2025.8.3
charset-normalizer==3.4.3
colorama==0.4.6
distro==1.9.0
Django==5.1.4
django-appconf==1.1.0
django-cors-headers==4.6.0
django-select2==8.4.1
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
iniconfig==2.0.0
jiter==0.10.0
mysqlclient==2.2.7
openai==1.107.2
packaging==24.1
pillow==11.1.0
pluggy==1.5.0
psycopg2==2.9.10
psycopg2-binary==2.9.10
pydantic==2.11.9
pydantic_core==2.33.2
PyJWT==2.9.0
pytest==8.3.3
requests==2.32.5
rest-framework-simplejwt==0.0.2
sniffio==1.3.1
sqlparse==0.5.3
stripe==12.5.1
tqdm==4.67.1
typing-inspection==0.4.1
typing_extensions==4.15.0
tzdata==2024.2
urllib3==2.5.0
rest_framework_nested
openai
stripe
